update file webdb_sederhana

+++ webdb_sederhana/app.js
@@ -16,7 +16,8 @@ const db = new sqlite3.Database("database.db");
 db.run(`
   CREATE TABLE IF NOT EXISTS items (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
-    nama TEXT
+    nama TEXT,
+    kelas TEXT
   )
 `);
 
@@ -26,15 +27,16 @@ app.get("/", (req, res) => {
     let html = `
     <html>
     <head>
-      <title>Data Siswa Kelas XI</title>
+      <title>Data Siswa </title>
       <link rel="stylesheet" href="/style.css">
     </head>
     <body>
       <div class="container">
-        <h1>Data Siswa Kelas XI</h1>
+        <h1>Data Siswa </h1>
 
         <form method="POST" action="/tambah">
           <input type="text" name="nama" placeholder="Masukkan nama" required>
+          <input type="text" name="kelas" placeholder="Masukkan kelas" required>
           <button type="submit">Tambah</button>
         </form>
 
@@ -44,8 +46,12 @@ app.get("/", (req, res) => {
     rows.forEach((item) => {
       html += `
         <li>
-          ${item.nama}
-          <a href="/hapus/${item.id}">Hapus</a>
+          <span class="nama">${item.nama}</span>
+          <span class="kelas">${item.kelas}</span>
+          <div class="aksi">
+            <a href="/edit/${item.id}" class="edit">Edit</a>
+            <a href="/hapus/${item.id}" class="hapus" onclick="return confirm('Yakin hapus?')">Hapus</a>
+          </div>
         </li>
       `;
     });
@@ -61,10 +67,51 @@ app.get("/", (req, res) => {
   });
 });
 
-// tambah
+// tambah nama
 app.post("/tambah", (req, res) => {
+  const { nama, kelas } = req.body;
+  db.run("INSERT INTO items (nama, kelas) VALUES (?, ?)", [nama, kelas]);
+  res.redirect("/");
+});
+
+
+app.get("/edit/:id", (req, res) => {
+  const { id } = req.params;
+
+  db.get("SELECT * FROM items WHERE id = ?", [id], (err, row) => {
+    const html = `
+    <html>
+    <head>
+      <title>Edit Data</title>
+      <link rel="stylesheet" href="/style.css">
+    </head>
+    <body>
+      <div class="container">
+        <h1>Edit Nama</h1>
+
+        <form method="POST" action="/update/${row.id}">
+          <input type="text" name="nama" value="${row.nama}" required>
+          <input type="text" name="kelas" value="${row.kelas}" required>
+          <button type="submit">Update</button>
+        </form>
+
+        <a href="/">Kembali</a>
+      </div>
+    </body>
+    </html>
+    `;
+
+    res.send(html);
+  });
+});
+
+// update data
+app.post("/update/:id", (req, res) => {
+  const { id } = req.params;
   const { nama } = req.body;
-  db.run("INSERT INTO items (nama) VALUES (?)", [nama]);
+  const { kelas } = req.body;
+
+  db.run("UPDATE items SET nama = ?, kelas = ? WHERE id = ?", [nama, kelas, id]);
   res.redirect("/");
 });
